UEH_BigData_Final/requirements.txt
fastf1>=3.0.0
pyspark>=3.3.0
mlflow>=1.30.0
pandas>=1.3.0
python-dotenv>=0.19.0
loguru
mkdocs
pip
python-dotenv
ruff
requests
numpy
mllib
retry
jsonschema
functools
matplotlib
tqdm
typer
-e .
